seperated logic from dags
